Add BSDStats package.
From their webpage:

The mission of this site is to try and come up with semi-accurate numbers for both advocacy, and marketing, of the *BSD Operating Systems.

More specifically, we are trying to prove to the various vendors out there that *BSD should be viewed as a serious Operating System, and not just a hobbiest system, for support (ie. hardware drivers) purposes.

Primarily started to provide numbers for FreeBSD deployments, we are designing the system so that NetBSD, OpenBSD and DragonFlyBSD can participate as well.

The script used for reporting can be found in the FreeBSD ports tree under sysutils/bsdstats.

If you have any questions, or suggestions, please email scrappy@freebsd.org