Fixing duplicate IP/Range entries
Because all three cases (CIDR, IP Range, and Single IP Address) are always
tested, it is possible that more than one will have a positive match.




Examples:
172.16.0.0/12 matches both for CIDR and Individual IP Address
169.254.0.0-169.254.255.255 matches for both Address Range and Individual IP




By doing if, elseif, elseif instead of if, if, if testing, the later tests
will only be performed if the former tests fail. Because the Individual IP
Address test will return a result, even for CIDRs and IP Ranges, that test
has been moved to the end of the elseif list.