Improve cron rcfile stop processing
If cron.pid did not exist, an error was emitted - that could happen when a stop is done at package install, before it was ever started. If cron is stopped, cron.pid was left behind. Another attempt to stop would emit an error about the pid in cron.pid no existing.
This extra checking avoids any error messages if stop is called before start, or stop is called twice.