Fix urlfilter_enable typo
I noticed this typo while looking at other code. I guess that when people have set the urlfilter_enable checkbox, it has not been actually implementing it all this time. Maybe nobody actually uses this option?
The question is, what if someone has checked this box and now the code is fixed and it suddenly starts working? Will that make a problem for anyone?
Also noticed a repeated unset command that I fixed up.