bandwidthd - check for interface subnet also specified by user
Check for the case where a user specifies the interface subnet to listen on in the "subnet" box, which is actually not required. In this case, do not put the subnet into the config twice. It fixes a GUI data entry confusion that is reported from time to time in the forum, e.g.
http://forum.pfsense.org/index.php/topic,49165.msg320801.html#msg320801
http://forum.pfsense.org/index.php/topic,66354.0.html