bandwidthd for 2.2 on nanoBSD
I installed bandwidthd on a 2.2 nanoBSD Alix2D13. The new path to the executable has an extra "/local" in it, and it needs an LD_LIBRARY_PATH to know where to find some of its libraries.
With this, it installs, runs and comes up again after a reboot.
These changes should all be backward-compatible and make no difference to what happens when installed on older 1.2 2.0 and 2.1 series systems.