Change way of stop openbgpd and add status on rc file bgpd.sh
The bgpd contains two linked process with him, In the rc file openbgpd is stoped by an killall command,  when the command kill the session engine, he automatically stop the another processes, but killall command try to stop the another processes too, generating an error on system.log file.

Trying to stop the session engine by the 'kill -15' command, and to verify if another processes is running before to use the killall command can be a better way to stop the bgpd.

Add status option in the rc file can help to see if the service is running.''