OpenVPN export fix index typo
The $server array here has and "index" key, not an "sindex" key.
Actually nothing was broken by this. When the user switches server selections, server_changed() is called. That uses JavaScript and selectedIndex to work out which entry the user selected. selectedIndex is a zero-based number. It matches the zero-based index of servers, which was built from the zero-based indexes of $ras_server array, which was built from data in the config. So by good luck or good management, this array index name typo is not actually a bug.
I noticed it while investigating forum https://forum.pfsense.org/index.php?topic=86263.0