Re-add the vnstat symlinks performed during package install
Note that this symlink would have originally failed as well because the package installs a vnstat.conf at that location, and ln without -f will refuse to overwrite an existing file.
By installing the symlink at install-time, the user is not required to press "Save" once in order for vnstat to begin collecting data.