pfBlockerNG - Improve Alerts Tab (Speed/Efficiency)
A significant change to the Alerts Tab. This was previously using the
pfSense conv_log_filter function which for this purpose was too bulky.
Added a new function conv_log_filter_lite to improve the process.

75% improvement in Processing Time/CPU Load ( even with alerts counts
>500 )

Improved the Javascript function so that its not called when not
required. This was spinning up additional php-fpm processes.

Changed !empty to !isset for Deny/Permit/Match Alert Counts. Previously
could not enter a "0" value with !empty.