Update apcupsd_mail.php
In Pfsense version 2.2 the variable/key $config['notifications']['smtp']['ssl'] is set with an empty string value when the parameter 'Enable SMTP over SSL/TLS' is checked in the webgui (and the key not exists when the checkbox is off). The check in this file is for the value 'checked', but is not working and ever happening, thats why I modify this to get it work with Gmail.