BIND - add support for multiple views per zone configuration
Often a zone configuration can be identical for LAN/local and WAN/pulic views, especially when using NAT reflection.
This change allows those zones to only be entered once in the configuration and then applied to both the local and public views.

The change converts the zone's 'view' field from a single selection to an array multiple selection.  Since an old value of 'local' is the same as an array value with only one element for 'local', this change is backwards compatible with existing configurations.  For that reason, I'm leaving the field name as 'view' rather than changing it to 'views'

When the zones are being added to the view clause, it will now check to see if the view is in the zone's view list rather than checking to see if it is a simple match before including the zone in the view.  By setting the old varible $zoneview to the matching view from the list, the existing code needs very little change.