OpenVPN Client Export code style
Since doctornoktor did a bunch of this, I thought that I might as well
review it also. It seems good to make the that are "internally"
maintained conform to the code style guide.
I bumped the version so it will be easy for people to see exactly what
version of code they are running, even though there is no functional
change here.