Redmine #5727 TFTP protect against empty stuff
1) If the user presses Save when there was no interface selected, and
the user has still not selected an interface, then do not attempt to
unset() tftpdinterface in the config when it is already unset.
2) If the user does backup/restore or other similar stuff when the
tftpboot directory has no files in it, then do not do chmod of the
(non-existent) files. This avoids a "No such file or directory" message
being logged.