BUGFIX: named / bind : resolve journal errors with ddns when service restarts
This fixes a bug with named when dynamic dns is enabled and zones are not started due to journal errors.  Existing code does a "killall" which does not allow named to flush the zone and journals to disk.  This results in zone files files that are not in sync causing named to not load the zone and for name resolution to fail.

* Fix: rc_stop() to use the recommended method for shutdown.
* Fix: no longer allows multiple instances to be accidentally started
* Add: rc_start detects and clear sthe journal when journal errors occur during startup.

These errors are critical for users running dynamic dns.

Recommend adding this to 2.2.