Table BGColor
the table is unreadable with current color settings and pfSense portal in dark theme.

Here are the colors I've tested: https://snag.gy/VFxeiK.jpg
#919191 is the line with "[blk_BL_automobile_bikes]"
Note that the word "access" at the end of the line is black to reflect light theme colors.

The lowest row, just above the developer tools, is the original presentation. The text is unreadable.